chore(plugin-hooks-fixes): - Updated error handling in source hooks. - Moved error handling related functions to error module. - Added copyright notices in new files.
